Explicitly set random seed for shuffles
Now the random generator used for shuffling sections (when that is enabled, as
it is by default) a random seed is set explicitly to make sure that shuffles are
predictable. Using the pangamebook filter twice on the same input-file to
generate two different output formats (e.g. PDF and EPUB of the same gamebook)
should now always create two output files with sections in the same order.

It is possible to set the random seed to any number using the gamebook-randomseed
metadata. That should normally not be needed.