Added new method of shuffling sections
Default now is to try to spread out sections
so that those that are adjacent in the
input document ends up N sections apart
in the output. The meta-data value
gamebook-gap is used to set what N
to use. If less than 1 the old
random shuffle will be used instead.

This commit also fixes a number of
bugs, including sometimes the last
sections of a story not being included
at all in the output.